first comit

+++ database/migrations/2025_10_17_194548_create_deposits_table.php
@@ -7,6 +7,9 @@ return new class extends Migration
 {
     public function up()
     {
+        if (Schema::hasTable('deposits')) {
+            return;
+        }
         Schema::create('deposits', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
